+++ tests/e2e/cases/array_prototype_value/in.ts
@@ -23,6 +23,7 @@ const staticValueOfDescriptor: any = Object.getOwnPropertyDescriptor(Array.proto
 console.log("static string descriptors:", staticMapDescriptor.value === proto.map, staticMapDescriptor.writable, staticMapDescriptor.enumerable, staticMapDescriptor.configurable, staticConstructorDescriptor.value === Array, staticConstructorDescriptor.writable, staticConstructorDescriptor.enumerable, staticConstructorDescriptor.configurable, staticValueOfDescriptor === undefined);
 console.log("own checks:", Object.hasOwn(Array.prototype, "map"), Object.hasOwn(Array.prototype, "valueOf"), proto.hasOwnProperty("constructor"), proto.propertyIsEnumerable("map"), Reflect.has(Array.prototype, "map"), Reflect.has(Array.prototype, "valueOf"));
 console.log("symbol checks:", Object.hasOwn(proto, Symbol.iterator), Object.hasOwn(proto, Symbol.unscopables), Object.hasOwn([], Symbol.iterator), Reflect.has(proto, Symbol.iterator), Reflect.has([], Symbol.iterator), Reflect.has(Object.prototype, Symbol.iterator));
+console.log("symbol proto methods:", proto.hasOwnProperty(Symbol.iterator), proto.propertyIsEnumerable(Symbol.iterator), Object.prototype.hasOwnProperty.call(proto, Symbol.unscopables), Object.prototype.propertyIsEnumerable.call(proto, Symbol.unscopables));
 const lengthDescriptor: any = Reflect.getOwnPropertyDescriptor(proto, "length");
 console.log("length descriptor:", lengthDescriptor.value, lengthDescriptor.writable, lengthDescriptor.enumerable, lengthDescriptor.configurable, Reflect.set(proto, "length", 1), proto.length);
 const map: any = proto.map;